fix(ts): move parseOptions to ts and adjust json import
