+++ lib/parseOptions.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import { resolve } from 'path'
 import format from 'date-fns/format'
 
-import { version } from '../package'
+import { version } from '../package.json'
 import { getHeadersConfig } from './utils/headers'
 import { proxyStringToObject, agentFromProxy } from 'contentful-batch-libs/dist/proxy'
 import addSequenceHeader from 'contentful-batch-libs/dist/add-sequence-header'
